requirements/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
anyjson==0.3.3
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.2
email-validator==2.1.0.post1
ephem==4.1.3
Flask==3.0.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.46.0
greenlet==3.0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
importlib-resources==6.1.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.3.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lml==0.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-memcached==1.58
pytz==2023.3.post1
PyYAML==5.4.1
redis==5.0.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.23
systemd-python==234
texttable==1.7.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.1
zipp==1.0.0
